# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Jinlian GYC API
  version: 4.3.0
  description: |
    金链·GYC 最新接口规范。
    当前版本关键约束：
    - UID 是业务主标识
    - 充值创建后自动到账，无人工审核
    - 提现仅交易时间可发起，提交即冻结并扣减暂定资产/总资产，确认完成不得二次扣款
    - 买入、卖出、提现统一手续费 0.1%
    - 点对点支付按“先增值后本金”扣款，收款方按 1‰ 收费（amount < 100 不收）
    - 后台支持手工转账/补款，实时改账并保留 ledger/audit/hash 留痕
    - 交易支持买入和卖出，自动撮合
    - 交易时段自动同步上金所官方开休盘，无需人工配置固定时段
    - 撮合规则必须写入代码，遵循价格优先、时间优先
    - 哈希使用 SHA-256，输出固定 64 位，并同步半公开金链
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /admin/funds/withdrawals/{orderId}/approve:
    post:
      tags: [AdminFunds]
      summary: 提现审核通过并进入转账处理中
      description: 同一订单只允许成功处理一次；重复点击会返回已处理结果或 409 冲突，并保留审计留痕。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWithdrawOrder'
        '409':
          description: 提现审核处理中或订单状态已变化
...
  /admin/funds/withdrawals/{orderId}/reject:
    post:
      tags: [AdminFunds]
      summary: 驳回提现并解冻回补资产
      description: 同一订单只允许成功驳回一次；重复点击会返回已处理结果或 409 冲突，并在事务内完成回补与留痕。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWithdrawOrder'
        '409':
          description: 提现驳回处理中或订单状态已变化
  /admin/funds/withdrawals/{orderId}/confirm-completed:
    post:
      tags: [AdminFunds]
      summary: 线下转账后确认提现完成
      description: 只允许更新订单为 completed，不允许二次扣减资产；重复点击会返回已处理结果或 409 冲突。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWithdrawOrder'
        '409':
          description: 提现完成确认处理中或订单状态已变化
  /admin/funds/manual-transfer:
    post:
      tags: [AdminFunds]
      summary: 后台手工转账
      description: 实时调整 tentativeAsset 和 totalAsset，并同步写入 ledger、audit、hash；需要手工转账权限。支持 `Idempotency-Key` 防止后台双击重复记账。
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: 幂等键。相同键的重复请求只允许一次成功入账，其余返回已处理结果或 409。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseManualFundAction'
        '409':
          description: 资金调整处理中或余额状态已变化
  /admin/funds/manual-adjust:
    post:
      tags: [AdminFunds]
      summary: 后台手工补款/冲正
      description: 实时调整 tentativeAsset 和 totalAsset，并同步写入 ledger、audit、hash；需要手工补款权限。支持 `Idempotency-Key` 防止后台双击重复记账。
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: 幂等键。相同键的重复请求只允许一次成功入账，其余返回已处理结果或 409。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseManualFundAction'
        '409':
          description: 资金调整处理中或余额状态已变化
...
components:
  schemas:
...
    WithdrawOrder:
      type: object
      properties:
        orderId:
          type: string
        queueNo:
          type: integer
          description: 提现队列号，按提交时间递增
        uid:
          type: integer
        amount:
          type: number
        feeRate:
          type: number
          example: 0.001
        feeAmount:
          type: number
        netAmount:
          type: number
        frozenAmount:
          type: number
        status:
          type: string
          enum: [pending_review, transfer_processing, completed, rejected]
        submittedAt:
          type: string
          format: date-time
        freezeAppliedAt:
          type: string
          format: date-time
        confirmCompletedAt:
          type: string
          format: date-time
        alertStatus:
          type: string
          enum: [ringing, muted, stopped]
        alertStoppedAt:
          type: string
          format: date-time
        alertStopReason:
          type: string
          enum: [open_detail, manual_mute]
        wechatQrCodeUrl:
          type: string
        alipayQrCodeUrl:
          type: string
        bankCardNo:
          type: string
        traceId:
          type: string
    ManualFundAction:
      type: object
      properties:
        uid:
          type: string
        traceId:
          type: string
        amount:
          type: number
        direction:
          type: string
          enum: [increase, decrease]
        reason:
          type: string
        latestBalance:
          type: object
          properties:
            tentativeAsset:
              type: number
            cashAsset:
              type: number
            totalAsset:
              type: number
            withdrawFrozenAmount:
              type: number
...
    ApiResponseWithdrawOrder:
      type: object
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/WithdrawOrder'
    ApiResponseManualFundAction:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/ManualFundAction'
...